simplify vhus policy
